pyproject.toml
[tool.poetry]
name = "sw-project"
version = "0.1.0"
description = ""
authors = ["abdallah"]
readme = "README.md"
packages = [{include = "sw_project"}]

[tool.poetry.dependencies]
python = "^3.8"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "39.0.0"
django = "4.1.5"
django-rest-knox = "4.2.0"
djangorestframework = "3.14.0"
drf-yasg = "1.21.4"
idna = "3.4"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.2"
markupsafe = "2.1.1"
packaging = "22.0"
pycparser = "2.21"
pytz = "2022.7"
requests = "2.28.1"
ruamel-yaml = "0.17.21"
ruamel-yaml-clib = "0.2.7"
sqlparse = "0.4.3"
tzdata = "2022.7"
uritemplate = "4.1.1"
urllib3 = "1.26.13"
asgiref = "3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
